services/inference/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
prometheus-client==0.20.0
pydantic==2.8.2
joblib==1.4.2
numpy==2.0.1
pandas==2.2.2
scikit-learn==1.5.1
psycopg2-binary==2.9.9
